Moved exceptions to shared package
