--- backend/src/main/java/pl/lodz/p/it/opinioncollector/exceptions/EmailAlreadyRegisteredException.java
@@ -1,4 +1,0 @@
-package pl.lodz.p.it.opinioncollector.exceptions;
-
-public class EmailAlreadyRegisteredException extends RuntimeException {
-}
