--- backend/src/main/java/pl/lodz/p/it/opinioncollector/exceptions/PasswordNotMatchesException.java
@@ -1,8 +1,0 @@
-package pl.lodz.p.it.opinioncollector.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class PasswordNotMatchesException extends Exception {
-}
