+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/exceptions/user/EmailAlreadyRegisteredException.java
@@ -1,0 +1,6 @@
+package pl.lodz.p.it.opinioncollector.exceptions.user;
+
+import pl.lodz.p.it.opinioncollector.exceptions.BaseApplicationException;
+
+public class EmailAlreadyRegisteredException extends BaseApplicationException {
+}
